clients

+++ app/Http/Controllers/ClientController.php
@@ -5,15 +5,14 @@ namespace App\Http\Controllers;
 use App\Helpers\TableParamsHelper;
 use App\Http\Requests\ClientRequest;
 use App\Models\Client;
+use App\Models\Locale;
+use App\Models\Module;
+use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 
 class ClientController extends Controller
 {
-    public function __construct()
-    {
-        $this->authorizeResource(Client::class, 'client');
-    }
 
     public function index(Request $request)
     {
@@ -35,42 +34,88 @@ class ClientController extends Controller
             TableParamsHelper::filterResponseAttributes($item, $request->get('attributes'), Client::class);
         }
 
-        return response($clients, 200);
+        return view('pages.clients.clients-list', [
+            'clients' => $clients,
+            'searchTerm' => $tableParams->search_term
+        ]);
     }
 
-    public function store(ClientRequest $request)
+    public function add()
     {
+        return view('pages.clients.single-client', [
+            'locales' => Locale::LOCALES_NAMES,
+            'modulesNames' => Module::MODULES_PAGE_NAMES,
+        ]);
+    }
+
+    public function edit(int $id)
+    {
+        $client = Client::where([
+            'id' => $id,
+        ])->first();
+
+        if (!$client) {
+            abort(404);
+        }
+        return view('pages.clients.single-client', [
+            'client' => $client,
+            'modulesNames' => Module::MODULES_PAGE_NAMES,
+            'locales' => Locale::LOCALES_NAMES
+        ]);
+    }
+
+    public function create(ClientRequest $request): RedirectResponse
+    {
+        $data = $request->all();
+        if (!count($data['available_locales'])) {
+            $data['available_locales'][] = $data['locale'];
+        }
         $client = new Client();
-        $client->fill($request->all());
+        $client->fill($data);
+        $client->slug = $data['slug'];
+        $client->modules = $data['modules'];
         $client->save();
 
-        return response(['id' => $client->id], 201);
+        return redirect()->route('clients');
     }
 
-    public function show(Client $client)
+    public function update(ClientRequest $request, int $id): RedirectResponse
     {
-        return response($client->toArray(), 200);
-    }
+        $client = Client::where([
+            'id' => $id,
+        ])->first();
 
-    public function update(ClientRequest $request, Client $client)
-    {
+        if (!$client) {
+            abort(404);
+        }
+        $data = $request->all();
+        $data['available_locales']=[];
+        if (!count($data['available_locales'])) {
+            $data['available_locales'][] = $data['locale'];
+        }
+        $client->fill($data);
         if (!Auth::user()->client->is_super_admin && $request->get('modules')) {
-            $data = $request->except('modules');
-        } else {
-            $data = $request->all();
+            $client->modules = $request->get('modules');
         }
 
-        $client->fill($data);
         $client->save();
-        return response(null, 200);
+        return redirect()->route('clients');
     }
 
-    public function destroy(Client $client)
+    public function destroy(int $id): RedirectResponse
     {
-        if (!Auth::user()->client->is_super_admin)
-            return response(null, 403);
+        $client = Client::where([
+            'id' => $id,
+        ])->first();
+
+        if (!$client) {
+            abort(404);
+        }
+        if (!Auth::user()->client->is_super_admin) {
+            abort(404);
+        }
 
         $client->delete();
-        return response(null, 200);
+        return redirect()->route('clients');
     }
 }
